remove redundant pwd
This allows scripts to be run in the project situated in a path containing spaces. Removing pwd is not an issue here, as the previous commit explicitly sets directory to project root and no commands in the scripts modify current directory and stay in it. Besides, relative paths are used everywhere in the scripts already