clean up build directory after script completion
In checks.sh, if ASAN finds problems, all execution is stopped, so it might not be possible to act in that case. This means that build directory is left out. Actually, might not be a problem, because in any case code should not contain these kinds of problems, especially not in done state